refactor: move resolveMergeReportProjects to blob.ts
